Extract general-purpose numba patterns into rarg-numba-patterns
Split out the following into a new rarg-numba-patterns PyPI package:

- intrinsics: load_data, accumulate_data, field_ptr, item_ptr,
  atomic_rmw_intrinsic, and their overload_method wrappers
- literals: Datum, DatumLiteral + subtypes, Schema, SchemaLiteral,
  LiteralStructRef, and all associated numba registrations
- concurrency: atomic spin-lock intrinsics (atomic_lock,
  lock_int_op, lock_array_op)
- tests: test_load_data, test_accumulate_data, test_pointers,
  Datum/LiteralStructRef test suite

radiomesh now re-exports through shim modules, so all existing
imports continue to work unchanged.
