+++ pyproject.toml
@@ -8,10 +8,11 @@ authors = [
 ]
 license = {text = "MIT"}
 readme = "README.md"
-requires-python = ">=3.10, <3.14"
+requires-python = ">=3.11, <3.14"
 dependencies = [
     "numpy>=2.0.0",
     "numba>=0.60.0",
+    "rarg-numba-patterns>=0.0.1",
 ]
 
 [dependency-groups]
